+++ JournalMonitor/JournalMonitor.cs
@@ -852,7 +852,24 @@ namespace EddiJournalMonitor
 
                                 if (!(from.StartsWith("$cmdr") || from.StartsWith("&")))
                                 {
-                                    // This is NPC speech; see if it's something that we can use
+                                    // This is NPC speech.  What's the source?
+                                    string source;
+                                    if (from.Contains("ShipName_Police"))
+                                    {
+                                        source = "Police";
+                                        from = getString(data, "From_Localised");
+                                    }
+                                    else if (from.Contains("npc_name_decorate"))
+                                    {
+                                        source = "NPC";
+                                        from = from.Replace("$npc_name_decorate:#name=", "").Replace(";", "");
+                                    }
+                                    else
+                                    {
+                                        source = "Station";
+                                    }
+                                    events.Add(new MessageReceivedEvent(timestamp, from, source, false, channel, getString(data, "Message_Localised")));
+                                    // See if we want to spawn a specific event as well
                                     if (message == "$STATION_NoFireZone_entered;")
                                     {
                                         events.Add(new StationNoFireZoneEnteredEvent(timestamp, false) { raw = line });
@@ -887,8 +904,23 @@ namespace EddiJournalMonitor
                                 }
                                 else
                                 {
-                                    from = from.Replace("$cmdr_decorate:#name=", "Commander ").Replace(";", "").Replace("&", "Commander ");
-                                    events.Add(new MessageReceivedEvent(timestamp, from, true, channel, message) { raw = line });
+                                    // Various sources
+                                    string source;
+                                    if (from.Contains("$RolePanel"))
+                                    {
+                                        source = "Crew member";
+                                        from = from.Replace("$RolePanel1_crew; $cmdr_decorate:#name=", "Crew member ");
+                                        from = from.Replace("$RolePanel1_unmanned; $cmdr_decorate:#name=", "Crew member ");
+                                        from = from.Replace("$RolePanel2_crew; $cmdr_decorate:#name=", "Crew member ");
+                                        from = from.Replace("$RolePanel2_unmanned; $cmdr_decorate:#name=", "Crew member ");
+                                    }
+                                    else
+                                    {
+                                        source = "Commander";
+                                        from = from.Replace("$cmdr_decorate:#name=", "").Replace("&", "");
+                                    }
+                                    from = from.Replace(";", "");
+                                    events.Add(new MessageReceivedEvent(timestamp, from, source, true, channel, message) { raw = line });
                                 }
                             }
                             handled = true;