+++ ChangeLog.md
@@ -3,6 +3,7 @@
     * Update 'Message received' event to include NPC messages.  Additional field 'Source' provides more details about the source of the message
   * Speech Responder
     * Add "Log" function to write information to EDDI's log.  This is an aid when debugging templates
+	* Fix issue where new templates might show up in custom personalities blank rather than with the contents of the default template
 	* Script changes:
 	  * 'Message received' - updated to only respond to player messages, and to use appropriate source
   * VoiceAttack Responder